% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_classes.R
\name{methyl_miss-class}
\alias{methyl_miss-class}
\title{methyl_miss Objects}
\description{
The \code{methyl_miss} class stores comprehensive information about missing data patterns
in a \code{\link{methyl_surro}} object. Objects of this class are created by the
\code{\link{methyl_miss}} function and provide detailed summaries of both partial and
complete missingness.
}
\section{Object Structure}{

A \code{methyl_miss} object is a list with class \code{"methyl_miss"} containing:
\describe{
\item{\code{missing_obs}}{A named numeric vector showing the proportion of missing
observations (0 to 1) for each probe that has partial missingness (i.e., missing in
some but not all samples). Names are CpG probe IDs.}
\item{\code{missing_probes}}{A character vector of CpG probe IDs that are completely
missing (i.e., missing in all samples).}
\item{\code{summary}}{A list containing overall summary statistics (see details below).}
}
}

\section{Summary Statistics}{

The \code{summary} component contains:
\describe{
\item{\code{total_probes}}{Total number of CpG probes in the methylation matrix.}
\item{\code{total_samples}}{Total number of samples in the methylation matrix.}
\item{\code{n_complete_probes}}{Number of probes with no missing values.}
\item{\code{n_missing_obs}}{Number of probes with partial missing data (some but not
all samples missing).}
\item{\code{n_missing_probes}}{Number of completely missing probes (all samples missing).}
\item{\code{overall_missing_rate}}{Proportion of all matrix values that are missing (0 to 1).}
\item{\code{missing_obs_rate}}{Proportion of probes with partial missing data (0 to 1).}
\item{\code{missing_probes_rate}}{Proportion of probes that are completely missing (0 to 1).}
\item{\code{complete_probes_rate}}{Proportion of probes with no missing values (0 to 1).}
}
}

\section{Creating methyl_miss Objects}{

Objects are created using \code{\link{methyl_miss}}:
\preformatted{
# Create methyl_surro object
my_surro <- surro_set(beta_matrix_miss, weights_vec, "Intercept")

# Assess missing data
miss_info <- methyl_miss(my_surro)
}
}

\section{Accessing Information}{

Components can be accessed using standard list notation:
\preformatted{
# View probes with partial missingness
miss_info$missing_obs

# View completely missing probes
miss_info$missing_probes

# Access summary statistics
miss_info$summary$overall_missing_rate
miss_info$summary$n_missing_probes
}
}

\section{Print Method}{

The package provides a \code{print} method that displays a formatted summary:
\preformatted{
print(miss_info)
}
This shows an organized view of missing data patterns and summary statistics.
}

\seealso{
\code{\link{methyl_miss}} for creating objects,
\code{\link{methyl_surro}} for the input object type,
\code{\link{reference_fill}} and \code{\link{impute_obs}} for addressing missing data
}
