% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_functions.R
\name{methylConversion}
\alias{methylConversion}
\alias{convert_m_to_beta}
\alias{convert_beta_to_m}
\title{Conversion Functions for M-Values and Beta Values}
\usage{
convert_m_to_beta(methyl)

convert_beta_to_m(methyl)
}
\arguments{
\item{methyl}{A matrix or data frame of methylation values. All values must be numeric and the matrix or data frame must have CpG sites as row names.}
}
\value{
For \code{convert_m_to_beta}, the function returns a matrix or data frame of beta values corresponding to the input M-values.

For \code{convert_beta_to_m}, the function returns a matrix or data frame of M-values corresponding to the input beta values.
}
\description{
These functions convert methylation M-values to beta values and vice versa.
}
\examples{
# Load Methylation M-Values Matrix
data(mval_matrix_comp, package = "MethylSurroGetR")
print(mval_matrix_comp)

# Convert M-values to Beta Values
beta_values <- convert_m_to_beta(mval_matrix_comp)
print(beta_values)

# Load Methylation Beta Matrix
data(beta_matrix_comp, package = "MethylSurroGetR")
print(beta_matrix_comp)

# Convert Beta Values to M-values
m_values <- convert_beta_to_m(beta_matrix_comp)
print(m_values)

}
