% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{mval_matrix_miss}
\alias{mval_matrix_miss}
\title{M-Values Matrix with Missing Values}
\format{
A numeric matrix with 15 rows (CpG probes) and 5 columns (samples):
\describe{
\item{Rows}{CpG probe identifiers matching other methylation matrices}
\item{Columns}{Sample identifiers matching other methylation matrices}
}
}
\source{
Converted from \code{\link{beta_matrix_miss}} using \code{convert_beta_to_m()}
}
\usage{
mval_matrix_miss
}
\description{
A 15 x 5 matrix of methylation M-values with the same missing values.
}
\examples{
data(mval_matrix_miss)
str(mval_matrix_miss)

}
\seealso{
\code{\link{mval_matrix_comp}}, \code{\link{beta_matrix_miss}}, \code{\link{convert_beta_to_m}}
}
\keyword{datasets}
