% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_obs.R
\name{print.imputation_stats}
\alias{print.imputation_stats}
\title{Print Method for imputation_stats Objects}
\usage{
\method{print}{imputation_stats}(x, ...)
}
\arguments{
\item{x}{An object of class \code{imputation_stats}.}

\item{...}{Additional arguments passed to print methods.}
}
\description{
Print Method for imputation_stats Objects
}
