% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_fill.R
\name{print.reference_fill_stats}
\alias{print.reference_fill_stats}
\title{Print Method for reference_fill_stats Objects}
\usage{
\method{print}{reference_fill_stats}(x, ...)
}
\arguments{
\item{x}{An object of class \code{reference_fill_stats}.}

\item{...}{Additional arguments passed to print methods.}
}
\description{
Print Method for reference_fill_stats Objects
}
