% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{ref_mat}
\alias{ref_mat}
\title{Reference Beta Values Matrix}
\format{
A data frame with 20 rows and 2 columns:
\describe{
\item{mean}{Mean beta value across samples}
\item{median}{Median beta value across samples}
}
}
\description{
A matrix containing the mean and median beta values for 20 reference probes.
}
\examples{
data(ref_mat)
str(ref_mat)
}
