% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{ref_vec_mean}
\alias{ref_vec_mean}
\title{Reference Beta Values Vector of Mean Values}
\format{
A named vector with 20 items.
}
\description{
A named vector containing the mean for 20 reference probes.
}
\examples{
data(ref_vec_mean)
str(ref_vec_mean)
}
