% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{ref_vec_median}
\alias{ref_vec_median}
\title{Reference Beta Values Vector of Median Values}
\format{
A named vector with 20 items.
}
\description{
A named vector containing the median for 20 reference probes.
}
\examples{
data(ref_vec_median)
str(ref_vec_median)
}
