% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{wts_vec_lin}
\alias{wts_vec_lin}
\title{Linear Regression Weights Vector}
\format{
A named vector with 10 elements.
}
\description{
A named vector of weights for the linear regression model.
}
\examples{
data(wts_vec_lin)
str(wts_vec_lin)
}
