% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{wts_vec_prb}
\alias{wts_vec_prb}
\title{Logistic Regression Weights Vector}
\format{
A named vector with 10 elements.
}
\description{
A named vector of weights for the logistic (probability) regression model.
}
\examples{
data(wts_vec_prb)
str(wts_vec_prb)
}
