% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{beta_matrix_miss}
\alias{beta_matrix_miss}
\title{Beta Matrix with Missing Values}
\format{
A matrix with 15 rows and 5 columns.
}
\description{
A 15 x 5 matrix of methylation beta values for sampled probes with some missing data.
(Probes are in rows and samples are in columns.)
}
\examples{
data(beta_matrix_miss)
str(beta_matrix_miss)
}
