% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_miss.R
\name{methyl_miss}
\alias{methyl_miss}
\title{Summarize Missing Data in methyl_surro Object}
\usage{
methyl_miss(methyl_surro)
}
\arguments{
\item{methyl_surro}{An object of class \code{methyl_surro}.}
}
\value{
A list with two elements:
\describe{
\item{missing_cases}{A named numeric vector with the proportion of missing cases for each probe where this value is greater than 0.}
\item{missing_probes}{A character vector of probes that are missing in all cases.}
}
}
\description{
This function summarizes the missing data in a methyl_surro object.
It identifies probes with partial missing data (missing cases) and probes missing
from all samples (missing probes).
}
\examples{
# Load Methylation Beta Matrix
data(beta_matrix_miss, package = "MethylSurroGetR")

# Load Weights from a Named Vector
data(wts_vec_lin, package = "MethylSurroGetR")

# Build the methyl_surro Object
surrogate <- surro_set(methyl = beta_matrix_miss,
                       weights = wts_vec_lin,
                       intercept = "Intercept")

# Summarizing Missing Values
missing_summary <- methyl_miss(methyl_surro = surrogate)
print(missing_summary)

}
