% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{methyl_surro_comp}
\alias{methyl_surro_comp}
\title{Constructed \code{methyl_surro} Object with Complete Data}
\format{
An object of class \code{methyl_surro}, containing the following components:
\describe{
\item{methyl}{A matrix of methylation data with CpG sites as row names. Some values are missing, simulating incomplete data scenarios.}
\item{weights}{A named numeric vector representing the weights for the CpG sites, excluding the intercept.}
\item{intercept}{The numeric value of the intercept term, or \code{NULL} if not applicable.}
}
}
\description{
A \code{methyl_surro} object constructed from \code{beta_matrix_comp} and \code{wts_vec_lin}.
}
\examples{
data(methyl_surro_comp)
str(methyl_surro_comp)
}
