% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{mval_matrix_comp}
\alias{mval_matrix_comp}
\title{Complete M-Values Matrix}
\format{
A numeric matrix with 15 rows (CpG probes) and 5 columns (samples):
\describe{
\item{Rows}{CpG probe identifiers matching \code{\link{beta_matrix_comp}}}
\item{Columns}{Sample identifiers matching \code{\link{beta_matrix_comp}}}
}
}
\source{
Converted from \code{\link{beta_matrix_comp}} using \code{convert_beta_to_m()}
}
\usage{
mval_matrix_comp
}
\description{
A 15 x 5 matrix of methylation M-values representing complete data without missing observations.
}
\examples{
data(mval_matrix_comp)
str(mval_matrix_comp)

}
\seealso{
\code{\link{beta_matrix_comp}}, \code{\link{mval_matrix_miss}}, \code{\link{convert_beta_to_m}}, \code{\link{convert_m_to_beta}}
}
\keyword{datasets}
