% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{mval_matrix_miss}
\alias{mval_matrix_miss}
\title{M-Values Matrix with Missing Values}
\format{
A matrix with 15 rows and 5 columns.
}
\description{
A 15 x 5 matrix of methylation M-values with some missing data.
(Probes are in rows and samples are in columns.)
}
\examples{
data(mval_matrix_miss)
str(mval_matrix_miss)
}
