% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_fill.R
\name{reference_fill}
\alias{reference_fill}
\title{Fill Missing Probes in methyl_surro Object Using Reference Data}
\usage{
reference_fill(
  methyl_surro,
  reference,
  col_name = NULL,
  type = c("probes", "obs", "all")
)
}
\arguments{
\item{methyl_surro}{An object of class \code{methyl_surro}, containing a methylation matrix.}

\item{reference}{A named numeric vector or a matrix/data frame with row names, containing reference values for probes to impute missing data.
If a matrix or data frame is provided, specify the column using the \code{col_name} argument.}

\item{col_name}{Optional; character string specifying the column to use if \code{reference} is a matrix or data frame.}

\item{type}{A character string specifying the scope for filling missing data:
\itemize{
\item \code{"probes"}: Fill in only missing probes.
\item \code{"obs"}: Fill in only missing observations
\item \code{"all"}: Fill in missing probes and observations
}}
}
\value{
A \code{methyl_surro} object with its methylation matrix updated by filling in the specified missing probes based on the reference data.
}
\description{
This function fills in missing probes in the methylation matrix of a \code{methyl_surro} object based on external reference data.
Users can specify if imputation should consider only fully missing rows, partially missing rows, or the entire dataset.
}
\examples{
# Load the sample data
data(methyl_surro_miss)
data(reference)

# Apply reference filling using a specific column of the reference data
result <- reference_fill(methyl_surro = methyl_surro_miss,
                         reference = reference,
                         col_name = "mean",
                         type = "probes")

# Check the result after filling
print(result$methyl)

}
