% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surro_calc.R
\name{surro_calc}
\alias{surro_calc}
\title{Calculate Predicted Values from methyl_surro Object}
\usage{
surro_calc(methyl_surro, transform = c("linear", "count", "probability"))
}
\arguments{
\item{methyl_surro}{An object of class \code{methyl_surro}.}

\item{transform}{A character string specifying the transformation to apply. It can be \code{"linear"}, \code{"count"}, or \code{"probability"}.}
}
\value{
A named vector of predicted values with names corresponding to the sample names in the methylation matrix.
}
\description{
This function calculates predicted values based on the provided methyl_surro object.
If any probes (rows) are completely missing, their values are set to zero, and a count of these
probes is stored. Columns (samples) with any missing values are removed, and a count of these
is also recorded. The function aligns and multiplies the weights with the methylation matrix,
adds the intercept if specified, and applies the specified transformation to the results.
}
\examples{
# Load \code{methyl_surro} Object
data(beta_matrix_comp, package = "MethylSurroGetR")

# Generating Surrogate Estimates from Linear Model Weights
pred_lin <- surro_calc(methyl_surro = methyl_surro_comp,
                       transform = "linear")
print(pred_lin)

}
