% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surro_set.R
\name{surro_set}
\alias{surro_set}
\title{Create a methyl_surro Object}
\usage{
surro_set(methyl, weights, intercept = NULL)
}
\arguments{
\item{methyl}{A matrix of methylation data. All values must be numeric and the matrix
must have CpG sites as row names.}

\item{weights}{A named numeric vector or a single-column data frame or matrix
with CpGs as row names, representing the weights.}

\item{intercept}{An optional name of the intercept term within the weights. If provided,
the specified intercept will be separated from the weights and stored separately.}
}
\value{
An object of class \code{methyl_surro} containing the following components:
\describe{
\item{methyl}{The input methylation matrix, filtered to include only probes present in the weights,
and with additional empty rows added for weights not present in the methylation data.}
\item{weights}{A named numeric vector of weights.}
\item{intercept}{The numeric value of the intercept, if provided. \code{NULL} otherwise.}
}
}
\description{
This function takes a methylation matrix and a set of weights (provided as a named vector,
a single-column data frame, or a single-column matrix with row names) and returns an object
of class \code{methyl_surro}.
}
\examples{
# Load Methylation Beta Matrix
data(beta_matrix_comp, package = "MethylSurroGetR")
print(beta_matrix_comp)

# Specify Weights from a Named Vector
data(wts_vec_lin, package = "MethylSurroGetR")
print(wts_vec_lin)

# Build the methyl_surro Object
surrogate_vec <- surro_set(methyl = beta_matrix_comp,
                           weights = wts_vec_lin,
                           intercept = "Intercept")
print(surrogate_vec)

# Specify Weights from a Data Frame
data(wts_df, package = "MethylSurroGetR")
print(wts_df)

# Build the methyl_surro Object
surrogate_df <- surro_set(methyl = beta_matrix_comp,
                       weights = wts_df["wt_lin"],
                       intercept = "Intercept")
print(surrogate_df)

# Specify Weights from a Matrix
data(wts_mat, package = "MethylSurroGetR")
print(wts_mat)

# Build the methyl_surro Object
surrogate_mat <- surro_set(methyl = beta_matrix_comp,
                           weights = wts_mat[, "wt_lin"],
                           intercept = "Intercept")
print(surrogate_mat)

}
